refactor: 파일 위치 이동, 테스트 코드 엔드포인트 수정 (#4)
- s3 관련 파일 s3 디렉토리 밑으로 이동
- 테스트 코드 엔드포인트 "/categorys" -> "categories" 로 수정

Refs: #4
